requirements.txt
Cython==0.27.3
numpy==1.14.1
scipy==1.0.1
pandas==0.22.0
quadprog==0.1.6
scikit-learn==0.19.1
PyYAML==3.12
jsonschema==2.6.0
